requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.8.2
httpx==0.27.0
python-dotenv==0.21.0
pyyaml==6.0.1
requests==2.32.3
langchain==0.2.10
langchain-community==0.2.10
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pillow==10.4.0
groq==0.1.0
playwright==1.46.0
pytest==8.3.2
pytest-asyncio==0.23.8